Web/requirements.txt
Flask==2.2.2
Flask-Compress==1.12
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Jinja2==3.1.2
joblib==1.1.0
pytest==7.2.0
pep8==1.7.1
scikit-learn==1.0.2
SQLAlchemy==1.3.22
gunicorn==20.1.0
tensorflow==2.3.1
Pillow==9.3.0
flask_cors==3.0.10
protobuf==3.19.6


